evalconfig function
This is the most general function in the "eval" family, allowing the caller
to effectively define an entirely new HCL-based language and then decode
configuration files written in that language.

This uses the "hcldec spec" language I previously designed for a hcldec
command line tool, but instead uses it to produce a value in memory, which
the function then returns.